refactor(ui): harmonize account & transaction messaging
