+++ packages/extension/src/shared/messages/AccountMessage.ts
@@ -5,14 +5,13 @@ export type AccountMessage =
   | {
       type: "NEW_ACCOUNT_RES"
       data: {
-        status: "ok"
         txHash: string
         address: string
         account: WalletAccount
         accounts: WalletAccount[]
       }
     }
-  | { type: "NEW_ACCOUNT_REJ"; data: { status: "ko"; error: string } }
+  | { type: "NEW_ACCOUNT_REJ"; data: { error: string } }
   | { type: "GET_ACCOUNTS" }
   | { type: "GET_ACCOUNTS_RES"; data: WalletAccount[] }
   | { type: "CONNECT_ACCOUNT"; data: WalletAccount }
