+++ packages/extension/src/ui/features/accounts/accounts.service.ts
@@ -1,6 +1,5 @@
 import { ethers } from "ethers"
 
-import { sendMessage } from "../../../shared/messages"
 import { startSession } from "../../services/backgroundSessions"
 import { Account } from "./Account"
 
@@ -10,17 +9,6 @@ export const deployAccount = async (networkId: string, password?: string) => {
   }
 
   return Account.fromDeploy(networkId)
-}
-
-export const connectAccount = (account: Account) => {
-  sendMessage({
-    type: "CONNECT_ACCOUNT",
-    data: {
-      address: account.address,
-      network: account.network,
-      signer: account.signer,
-    },
-  })
 }
 
 const argentColorsArray = [
