+++ packages/extension/src/ui/features/onboarding/MigrationDisclaimerScreen.tsx
@@ -8,7 +8,8 @@ import { Button } from "../../components/Button"
 import { OpenInNewIcon } from "../../components/Icons/MuiIcons"
 import { H2, P } from "../../components/Typography"
 import { routes } from "../../routes"
-import { connectAccount, deployAccount } from "../accounts/accounts.service"
+import { connectAccount } from "../../services/backgroundAccounts"
+import { deployAccount } from "../accounts/accounts.service"
 import { useAccounts } from "../accounts/accounts.state"
 import { recover } from "../recovery/recovery.service"
