+++ packages/extension/src/ui/features/onboarding/NewWalletScreen.tsx
@@ -10,7 +10,8 @@ import { InputText } from "../../components/InputText"
 import { FormError, H2, P } from "../../components/Typography"
 import { routes } from "../../routes"
 import { analytics, usePageTracking } from "../../services/analytics"
-import { connectAccount, deployAccount } from "../accounts/accounts.service"
+import { connectAccount } from "../../services/backgroundAccounts"
+import { deployAccount } from "../accounts/accounts.service"
 import { useAccounts } from "../accounts/accounts.state"
 import { StickyGroup } from "../actions/ConfirmScreen"
 import { recover } from "../recovery/recovery.service"
